pyproject.toml
[tool.poetry]
name = "gpt3-blog-title-optimizer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-bigquery = "^3.11.3"
openai = "^0.27.8"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
